feat : 백엔드 프론트 연결 작업 (station insert, place insert, place list, station list구현 완료)

+++ back/app/src/main/java/com/sloway/app/place/service/station/StationService.java
@@ -46,7 +46,7 @@ public class StationService {
 
         //편의시설 정보 조회
         List<Long> amenityNos = dto.getFacilityList().stream()
-                .map(facility -> (long) facility.getAmenityNo())
+                .map(facility -> facility.getAmenityNo())
                 .collect(Collectors.toList());
         List<AmenityEntity> amenityEntities = amenityRepository.findAllByNoIn(amenityNos);
